% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex_de_spa.R
\docType{data}
\name{ex_de_spa}
\alias{ex_de_spa}
\title{The data set with coordinates for ex SPA}
\format{
data frame with 44 rows and 6 columns
}
\usage{
data(ex_de_spa)
}
\description{
The dataset contains coordinates for places that used to have SPA status  (Pol. sanatoria), but now are not considered as SPA anymore.
The subsequent columns in this data set are the following:
\itemize{
\item MIEJSCE - SPA's location
\item LECZENIE - SPA’s profile
\item LAT - SPA’s latitude
\item LON - SPA’s longitude
\item WOJEWODZTWO - SPA's province (voivodeship)
\item CATEGORY - SPA's category
}
}
\keyword{datasets}
