% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.all.dist.R
\name{find.all.dist}
\alias{find.all.dist}
\title{This function finds all distances between provided coordinates and listed object's coordinates}
\usage{
find.all.dist(touristic_data, lat_my, lon_my)
}
\arguments{
\item{The}{\code{find.all.dist()} function is called with data frame argument,
which contains a set of listed objects coordinates.}
}
\value{
The \code{find.all.dist()} function returns a data frame with two rows,
which contain two LAT LON coordinates with the longest distance.
}
\description{
The \code{find.all.dist()} function calculates all distances between the coordinates provided as input and all listed coordinates.
The function uses haversine function for distance calculations.
}
\details{
The \code{find.all.dist()} function is called with data frame argument
that should contain at least the following columns:
LAT, LON, LECZENIE, DATETIME, WOJEWODZTWO

The result of the \code{find.all.dist()} is a data frame with two rows
containing two AIS signal observations
and with the longest distance between two consecutive observations reported
in new column named DIST.
Consequently, the result contains the following columns:
LAT, LON, DIST, LECZENIE, WOJEWODZTWO

Please find below short description of AIS signal columns:

LAT - ship’s latitude

LON - ship’s longitude

DIST - calculated the longest distance between two consecutive observations

LECZENIE - touristic object’s name

WOJEWODZTWO - touristic object’s province (voivodeship)
}
\examples{
\dontrun{
find.all.dist(history_monuments)
}

}
\author{
Dariusz Ekonomiuk, \email{ekodarek@gmail.com}
}
