% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.max.dist.R
\name{find.max.dist}
\alias{find.max.dist}
\title{This function finds maximum distance between provided coordinates and listed object's coordinates}
\usage{
find.max.dist(touristic_data, lat_my, lon_my)
}
\arguments{
\item{The}{\code{find.max.dist()} function is called with data frame argument,
which contains a set of touristic attractions coordinates.}
}
\value{
The \code{find.max.dist()} function returns a data frame with two rows: the first row contains coordinates provided by user
and the second row contains coordinates for object with the longest distance from the first one.
}
\description{
The \code{find.max.dist()} function analyzes touristic attractions coordinates and identifies two records
whith the longest distance
}
\details{
The \code{find.max.dist()} function is called with data frame argument
that should contain at least the following columns:
LAT, LON, LECZENIE, DATETIME, WOJEWODZTWO

The result of the \code{find.max.dist()} is a data frame with two rows
containing two AIS signal observations
and with the longest distance between two consecutive observations reported
in new column named DIST.
Consequently, the result contains the following columns:
LAT, LON, DIST, LECZENIE, WOJEWODZTWO

Please find below short description of AIS signal columns:

LAT - ship’s latitude

LON - ship’s longitude

DIST - calculated the longest distance between two consecutive observations

LECZENIE - touristic object’s name

WOJEWODZTWO - touristic object’s province (voivodeship)
}
\examples{
\dontrun{
find.max.dist(history_monuments)
}

}
\author{
Dariusz Ekonomiuk, \email{ekodarek@gmail.com}
}
