% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.gui.R
\name{run.gui}
\alias{run.gui}
\title{Run GUI (shiny app) with leaflet map visualization.}
\usage{
run.gui()
}
\arguments{
\item{The}{\code{run.gui()} function is called without arguments.}
}
\description{
The \code{run.gui()} function executes shiny.
}
\details{
Run this function and you will observe visualization
of your location and the closest SPA or ex-SPA on leaflet map and the distance between these two points will be calculated.
}
\author{
Dariusz Ekonomiuk, \email{ekodarek@gmail.com}
}
