% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spa_poland.R
\docType{data}
\name{spa_poland}
\alias{spa_poland}
\title{The dataset with coordinates for Polish SPA}
\format{
data frame with 44 rows and 6 columns
}
\usage{
data(history_monuments)
}
\description{
The dataset contains coordinates for SPA in Poland (Pol. sanatoria).
The subsequent columns in this dataset are the following:
\itemize{
\item MIEJSCE - SPA's location
\item LECZENIE - SPA’s profile
\item LAT - SPA’s latitude
\item LON - SPA’s longitude
\item WOJEWODZTWO - SPA's province
\item CATEGORY - SPA's category
}
}
\keyword{datasets}
